Add Code of the Day PR flagging (#12)
A user wanted to be able to flag a certain PR thread so that he could later write a "Code of the Day" blog post about it. To avoid embarrassing anyone, it has to be implemented in a way that it's not obvious when someone's PR/thread has been flagged.

I added a button next to the "like comment" button that toggles whether the thread is flagged (individual comments are not flagged). When a thread is flagged the button icon and the tooltip changes. You only see the flag statuses that you have set. If other people flag a thread, you will not see it, and you can toggle your own flagged state for that thread independently.

The flagged thread states for a particular PR are persisted in a hidden property (NI.CodeOfTheDay) on that PR. These states will be fetched via a script. One limitation of this persistence approach is that completing a PR effectively "freezes" all COD flagging on it. You can't modify the properties on a completed PR.