Sort prs by added date (#15)
Before separating PR dashboard rows into different buckets (e.g. Drafts, Approved, etc), set their relative order based on when the user was (most recently) added to the review.

The most direct way I know of to get that information is by looking for the comment thread that announces when a reviewer is added to the PR. We were already querying the threads for PRs that the user had approved (in order to check for "notable activity"), but now we have to do this for all PRs the user is on. That will result in a performance hit, but it should be acceptable.