Disable PR file checkboxes now that AzDO has them
AzDO implemented them!

https://docs.microsoft.com/en-us/azure/devops/release-notes/2019/sprint-159-update#mark-files-as-reviewed-in-a-pull-request

Disabling them if I notice evidence that they are on the page, or if we're on the NI account, which I know has them. 

Note:Sometimes, the evidence is not there when I check, and if I'm wrong, you get two checkboxes. This is OK for now. I will fully disable checkboxes later this week.