Run the script at document.body, adding CSS as early as possible (#58)
In what appears to be a Chrome bug, if the scrollbar CSS is not available when it's first being rendered, the scrollbar does not restyle when CSS is added to it later.

So now, we apply the scrollbar CSS when the document body is first created, while keeping the old behavior of applying other modifications only after we've loaded the DOM.

I've also changed the throttling to apply only to trailing edges. Before, we ran the update function at the beginning and end of the throttling period. This is a wasteful use of CPU. Now, browing AzDO appears a little faster.