PR dashboard should not show multiple SEV tags on a PR (#79)
Instead of annotating the item with a tag for every linked bug, tag it at most once, for the highest severity (i.e. lowest number). The tooltip will have the title of the first bug with the highest severity. If there are more than one with the same severity, the title will be followed by " (and N other)".