Add a trophies section to the PR Overview page (#82)
### Justification
Add a trophies section to the right-hand pane on the Overview tab of a PR containing "Trophies" for notable PR numbers. This was a quirky feature of Review Board that I thought it would be fun to bring over to AzDo.
### Testing
Works for: Firefox and Chrome; light theme and dark theme; dev.azure.com/account and account.visualstudio.com.
Test PRs from NI: There are no valid milestone PRs that I could find, so I tested the logic by hard-coding a PR number. PRs 75057 and 70007 are valid for the fish trophy. 

### Known issues
Does not work for the "new PR experience" preview feature. I have version that partially works for the new PR view that I could finish out when that becomes the default view. This version does not cause any page or script errors for the new view - the code terminates early when it can't find the PR author from the same part of the HTML.

### Screenshots
Review Board version:
![review-board-trophies](https://user-images.githubusercontent.com/35350751/83075210-14288000-a039-11ea-99ea-3c04d4ec5241.png)
AzDo Userscript version:
![milestone-trophy-large](https://user-images.githubusercontent.com/35350751/83074155-5224a480-a037-11ea-96c3-cef0f5d2b63b.png)
![fish-trophy](https://user-images.githubusercontent.com/35350751/83074180-5f419380-a037-11ea-958e-91c3444b9de8.png)