Avoid error when approving in new experience (#92)
The bypass reminder resulted in a browser error when the user clicks the "Approve" button (in the new PR experience, only). Doing so inserts a vote icon (e.g. green circle with checkmark) to the left of the Approve button, and apparently the logic depended on certain DOM element relationships that I altered. To avoid this, I'm now leaving the vote button with the same parent as before. I'm instead pushing my changes down into the vote button itself (rather than wrapping the vote button in new elements).

Co-authored-by: Mert Akinc <mert.akinc@ni.com>