Fix to follower list not showing others (#127)
As it turns out, follower lists are private info to each user. Because of that, the follower listing on work items was only ever showing you, and not others -- well unless you were a super user (like me on the NI account).

To get around this, we need to perform the subscription query with the alwaysReturnBasicInformation query flag. This returns limited info on a subscription.

The limited info means we can't tell if the user is following the full work item or just state changes. But that's an acceptable limitation. We'll still list them in the follower list.

Fixes #123