Update help text for NI Binary Git Diff
The current help text says to contact Humberto. We now have instructions in the wiki, so I'm updating the help text to point to the wiki (without including the internal-only link). I tested that searching for that string in the NI Azure DevOps wiki will return the instructions we need as the first result.