Add a copy branch name button to Git repo pages (#170)
From a user:

This has frustrated me a few times. I want to git checkout some long branch name and I'm starting with a link that takes me to the branch in AzDO (e.g. this).

However, the place on this page that shows you what the branch name is does not allow you to select the text. I can right-click the drop-down element and click Inspect, and that will get me to the raw text in the markup fairly quickly, but it shouldn't have to be like that.